feat: add fab for creating board(#354)
