+++ src/widgets/board/ui/BoardList/BoardListClient.tsx
@@ -2,9 +2,9 @@
 
 import { useEffect, useState } from 'react';
 
-import Link from 'next/link';
-
 import { isGraduate, useUserAcademic } from '@/entities/user';
+
+import { Fab } from '@/shared';
 
 import { CustomBoard } from './CustomBoard';
 import { DefaultNoticeBoard } from './DefaultNoticeBoard';
@@ -22,14 +22,6 @@ export const BoardListClient = ({
 
   useEffect(() => setMounted(true), []);
   if (!mounted) return null;
-
-  const fabStyles =
-    'group fixed right-6 bottom-24 xl:right-80 xl:bottom-10 h-16 w-16 xl:h-24 xl:w-24 ' +
-    'rounded-[50px] bg-[#7AB6C1] text-white text-3xl font-normal shadow-lg ' +
-    'transition-all duration-200 ease-out will-change-transform ' +
-    'hover:-translate-y-0.5 hover:shadow-xl hover:bg-[#6AA3AD] active:scale-95 ' +
-    'focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 focus-visible:ring-[#7AB6C1] ' +
-    'motion-reduce:transition-none motion-reduce:hover:translate-y-0';
 
   const renderBoard = () => {
     if (roles.includes('ADMIN')) {
@@ -65,18 +57,7 @@ export const BoardListClient = ({
         <div className="flex flex-col items-center">{renderBoard()}</div>
       </div>
 
-      {!isGraduated && (
-        <Link href="/board/create" aria-label="게시글 작성">
-          <button type="button" title="게시글 작성" className={fabStyles}>
-            <span
-              aria-hidden
-              className="inline-block leading-none transition-transform duration-200 group-hover:scale-110"
-            >
-              +
-            </span>
-          </button>
-        </Link>
-      )}
+      {!isGraduated && <Fab href="/board/create" label="게시판 추가" />}
     </>
   );
 };
